requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.4
django-livereload-server==0.4
djangorestframework==3.14.0
psycopg2==2.9.7
pytz==2023.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
